Create {Mutable,}IdentifiableCollection
